move template helpers to format utils, refactor format utils from class to exported functions
