+++ app/assets/javascripts/dashboard/advanced_dataset/dataset_access_list_view.js
@@ -2,7 +2,7 @@
 /* eslint-disable jsx-a11y/href-no-hash */
 
 import * as React from "react";
-import { stringToColor } from "libs/template_helpers";
+import { stringToColor } from "libs/format_utils";
 import { getDatasetAccessList } from "admin/admin_rest_api";
 import type { APIDatasetType, APIUserType } from "admin/api_flow_types";
 import { Spin, Tag } from "antd";
